Enum sc_peerset::DropReason

pub enum DropReason {
    Unknown,
    Refused,
}

Reason for calling Peerset::dropped.

Variants

Unknown

Substream or connection has been closed for an unknown reason.

Refused

Substream or connection has been explicitly refused by the target. In other words, the peer doesn’t actually belong to this set.

This has the side effect of calling PeersetHandle::remove_from_peers_set.