Struct sc_peerset::Peerset

pub struct Peerset { /* fields omitted */ }

Side of the peer set manager owned by the network. In other words, the "receiving" side.

Implements the Stream trait and can be polled for messages. The Stream never ends and never errors.

Methods

impl Peerset

pub fn from_config(config: PeersetConfig) -> (Peerset, PeersetHandle)

Builds a new peerset from the given configuration.

pub fn incoming(&mut self, peer_id: PeerId, index: IncomingIndex)

Indicate that we received an incoming connection. Must be answered either with a corresponding Accept or Reject, except if we were already connected to this peer.

Note that this mechanism is orthogonal to Connect/Drop. Accepting an incoming connection implicitly means Connect, but incoming connections aren't cancelled by dropped.

pub fn dropped(&mut self, peer_id: PeerId)

Indicate that we dropped an active connection with a peer, or that we failed to connect.

Must only be called after the PSM has either generated a Connect message with this PeerId, or accepted an incoming connection with this PeerId.

pub fn discovered<I: IntoIterator<Item = PeerId>>(&mut self, peer_ids: I)

Adds discovered peer ids to the PSM.

Note: There is no equivalent "expired" message, meaning that it is the responsibility of the PSM to remove PeerIds that fail to dial too often.

pub fn report_peer(&mut self, peer_id: PeerId, score_diff: ReputationChange)

Reports an adjustment to the reputation of the given peer.

pub fn debug_info(&mut self) -> Value

Produces a JSON object containing the state of the peerset manager, for debugging purposes.

pub fn get_priority_group(&self, group_id: &str) -> Option<HashSet<PeerId>>

Returns priority group by id.