Crate sc_peerset

Peer Set Manager (PSM). Contains the strategy for choosing which nodes the network should be connected to.

Structs

IncomingIndex
Opaque identifier for an incoming connection. Allocated by the network.

PeerId
Identifier of a peer of the network.

Peerset
Side of the peer set manager owned by the network. In other words, the “receiving” side.

PeersetConfig
Configuration to pass when creating the peer set manager.

PeersetHandle
Shared handle to the peer set manager (PSM). Distributed around the code.

ReputationChange
Description of a reputation adjustment for a node.

Enums

Message
Message that can be sent by the peer set manager (PSM).