Struct sc_peerset::Peerset

pub struct Peerset { /* private fields */ }

Side of the peer set manager owned by the network. In other words, the “receiving” side.

Implements the Stream trait and can be polled for messages. The Stream never ends and never errors.

Implementations

impl Peerset

pub fn from_config(config: PeersetConfig) -> (Self, PeersetHandle)

Builds a new peerset from the given configuration.

pub fn reserved_peers(&self, set_id: SetId) -> impl Iterator<Item = &PeerId>

Returns the list of reserved peers.

pub fn add_to_peers_set(&mut self, set_id: SetId, peer_id: PeerId)

Adds a node to the given set. The peerset will, if possible and not already the case, try to connect to it.

Note: This has the same effect as PeersetHandle::add_to_peers_set.

pub fn incoming(&mut self, set_id: SetId, peer_id: PeerId, index: IncomingIndex)

Indicate that we received an incoming connection. Must be answered either with a corresponding Accept or Reject, except if we were already connected to this peer.

Note that this mechanism is orthogonal to Connect/Drop. Accepting an incoming connection implicitly means Connect, but incoming connections aren’t cancelled by dropped.

pub fn dropped(&mut self, set_id: SetId, peer_id: PeerId, reason: DropReason)

Indicate that we dropped an active connection with a peer, or that we failed to connect.

Must only be called after the PSM has either generated a Connect message with this PeerId, or accepted an incoming connection with this PeerId.

pub fn report_peer(&mut self, peer_id: PeerId, score_diff: ReputationChange)

Reports an adjustment to the reputation of the given peer.

pub fn debug_info(&mut self) -> Value

Produces a JSON object containing the state of the peerset manager, for debugging purposes.

pub fn num_discovered_peers(&self) -> usize

Returns the number of peers that we have discovered.